add territory response
